+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.19.1] - 2018-06-07
 ### Changed
+- Fixed bug in missing unpack when retrieving BCF record id.
 - Fixed bug in the filter iterator of BCF.
 
 ## [0.19.0] - 2018-06-01